functionsV2/tests/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
psycopg2-binary==2.9.9
flask==3.0.0
python-dotenv==1.0.0

